+++ windows_version_info.py
@@ -1,0 +1,30 @@
+VSVersionInfo(
+  ffi=FixedFileInfo(
+    filevers=(1, 1, 1, 0),
+    prodvers=(1, 1, 1, 0),
+    mask=0x3f,
+    flags=0x0,
+    OS=0x4,
+    fileType=0x1,
+    subtype=0x0,
+    date=(0, 0)
+  ),
+  kids=[
+    StringFileInfo([
+      StringTable(
+        u'040904B0',
+        [
+          StringStruct(u'CompanyName', u''),
+          StringStruct(u'FileDescription', u'Ultrawide mod for Hades'),
+          StringStruct(u'FileVersion', u'1.1.1'),
+          StringStruct(u'InternalName', u'hephaistos'),
+          StringStruct(u'LegalCopyright', u''),
+          StringStruct(u'OriginalFilename', u'hephaistos.exe'),
+          StringStruct(u'ProductName', u'Hephaistos'),
+          StringStruct(u'ProductVersion', u'1.1.1')
+        ]
+      )
+    ]),
+    VarFileInfo([VarStruct(u'Translation', [1033, 1200])])
+  ]
+)